a21s: numerous fixes
Added in more vendor blobs, fixed fingerprint sample removal issues, and included makefile for telephony parts, which were previously missing.

Please note that this is still a WIP. SELinux policies are still incomplete (thus builds need to be in permissive mode to function), and additional tests are needed to make sure that this is fully functional.

Fingerprint blob was pinned to solve a problem that occured when deleting enrolled fingerprint samples. This occured due to libbauthserver returning errors on LineageOS builds, and solution was achieved by hexpatching the lib with the following:

BEFORE:

53f68:54000341 b.ne 53fd0 <_ZN7android14FPBAuthService22removeEnrolledTemplateENS_7String8Ei@@Base+0x1c4>

53f74:540002e1 b.ne 53fd0 <_ZN7android14FPBAuthService22removeEnrolledTemplateENS_7String8Ei@@Base+0x1c4>

AFTER:

53f68:54000340 b.eq 53fd0 <_ZN7android14FPBAuthService22removeEnrolledTemplateENS_7String8Ei@@Base+0x1c4>

53f74:540002e0 b.eq 53fd0 <_ZN7android14FPBAuthService22removeEnrolledTemplateENS_7String8Ei@@Base+0x1c4>

41 03 00 54 68 2f 48 b9 1f 05 00 71 e1 02 00 54 => 40 03 00 54 68 2f 48 b9 1f 05 00 71 e0 02 00 54